requirements-full.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.4.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.0.1
transformers==4.30.2
